Move getFieldControlWithFallback to registry
