+++ src/utils/form-helper.ts
@@ -158,32 +158,6 @@ export function findConceptByReference(reference: string, concepts) {
   }
 }
 
-/**
- * Retrieves the appropriate field control for a question, considering missing concepts.
- * If the question is of type 'obs' and has a missing concept, it falls back to a disabled text input.
- * Otherwise, it retrieves the registered control based on the rendering specified in the question.
- * @param question - The FormField representing the question.
- * @returns The field control to be used for rendering the question.
- */
-export function getFieldControlWithFallback(question: FormField) {
-  // Check if the question has a missing concept
-  if (hasMissingConcept(question)) {
-    // If so, render a disabled text input
-    question.disabled = true;
-    question.isDisabled = true;
-    return getRegisteredControl('text');
-  }
-
-  // Retrieve the registered control based on the specified rendering
-  return getRegisteredControl(question.questionOptions.rendering);
-}
-
-export function hasMissingConcept(question: FormField) {
-  return (
-    question.type == 'obs' && !question.questionOptions.concept && question.questionOptions.rendering !== 'fixed-value'
-  );
-}
-
 export function scrollIntoView(viewId: string, shouldFocus: boolean = false) {
   const currentElement = document.getElementById(viewId);
   currentElement?.scrollIntoView({
